chg:usr:#87:Standardize syntax of DROP & SHOW. Add multi-tenant support (#90)
This PR has the following two main contributions:

1.Standardize syntax of SHOW & DDL. WHERE in both these DDLS is not standard. Moreover, the way WHERE is implemented is a security flaw. It can be used for SQL Injection. I have closed this whole. WHERE makes sense for SHOW. I have added support for LIKE which is easier & safer to implement given the current architecture. These changes have to be made to CREATE and ALTER. I'll take that up as part of #80

2.Support multi-tenancy. The main idea is to seed all DAO calls with ds_set_id. For e.g. check findByName or findById. Then a moved some code around and created Connection class to initialize and get DSSet.

+++ catalog/src/main/java/com/qubole/quark/catalog/db/dao/DataSourceDAO.java
@@ -53,7 +53,7 @@ public abstract class DataSourceDAO {
   public abstract int update(@BindBean("d") DataSource ds);
 
   @SqlUpdate("delete from data_sources where id = :id")
-  public abstract void delete(@Bind("id") int id);
+  public abstract void delete(@Bind("id") long id);
 
   public int insertJDBC(String name, String type, String url, long dsSetId,
                         String dataSourcetype, JdbcSourceDAO jdbcDao, String username,
